Register and Login API views and serializer are moved into their own files.
