+++ backend/main/api/loginapi.py
@@ -1,0 +1,14 @@
+from rest_framework.views import APIView
+from rest_framework.response import Response
+import status
+
+from ..serializers.loginserializer import LoginSerializer
+
+
+class Login(APIView):
+    @staticmethod
+    def post(request):
+        serializer = LoginSerializer(data=request.data)
+        if not serializer.is_valid():
+            return Response(serializer.errors, status.HTTP_400_BAD_REQUEST)
+        return Response(serializer.data, status.HTTP_200_OK)
